removing typescript, swapping to mjs, prettier, etc
