+++ CONTRIBUTING.md
@@ -1,6 +1,5 @@
-A lot of effort has been put into `faker.js` to create a useful and handy
-library. There are still a lot of things to be done, so all contributions are
-welcome! If you can make `faker.js` better, please read the following contribution guide.
+A lot of effort has been put into `Faker` to create a useful and handy
+library. There are still a lot of things to be done, so all contributions are welcome! If you can make `Faker` better, please read the following contribution guide.
 
 # Important
 
@@ -8,8 +7,7 @@ welcome! If you can make `faker.js` better, please read the following contributi
 
 ## Support
 
-`faker.js` relies on [commonJS](http://www.commonjs.org/) standard and supports both node.js and the
-browsers. Keep this in mind, when modifying and/or extending the sources.
+`Faker` relies on [commonJS](http://www.commonjs.org/) standard and supports both node.js and the browsers. Keep this in mind, when modifying and/or extending the sources.
 
 ## Automation
