+++ README.md
@@ -1,4 +1,4 @@
-# faker.js
+# Faker
 
 Generate massive amounts of fake data in the Browser and Node.js.
 
@@ -8,7 +8,7 @@ Generate massive amounts of fake data in the Browser and Node.js.
 
 This project was originally created and hosted at https://github.com/marak/Faker.js/ - however around 4th Jan, 2022 - the author decided to delete the repository (for unknown reasons).
 
-In the interest of the community, it has been decided that faker.js will continue to be maintained here and all help in its development will be appreciated.
+In the interest of the community, it has been decided that Faker will continue to be maintained here and all help in its development will be appreciated.
 
 ## Demo
 
@@ -269,7 +269,7 @@ var randomCard = faker.helpers.createCard(); // random contact card containing m
 
 ### Faker.fake()
 
-faker.js contains a super useful generator method `Faker.fake` for combining faker API methods using a mustache string format.
+Faker contains a super useful generator method `Faker.fake` for combining faker API methods using a mustache string format.
 
 **Example:**
 
@@ -283,7 +283,7 @@ This will interpolate the format string with the value of methods `name.lastName
 
 ## Localization
 
-As of version `v2.0.0` faker.js has support for multiple localities.
+As of version `v2.0.0` Faker has support for multiple localities.
 
 The default language locale is set to English.
 
@@ -343,7 +343,7 @@ faker.locale = 'de';
 
 ### Individual Localization Packages
 
-faker.js supports incremental loading of locales.
+Faker supports incremental loading of locales.
 
 By default, requiring `faker` will include _all_ locale data.
 
@@ -373,14 +373,16 @@ console.log(firstRandom === secondRandom);
 
 ## Documentation
 
-faker.js is currently in the process of migrating its documentation to Vitepress (the successor of Vuepress).
+Faker is currently in the process of migrating its documentation to Vitepress (the successor of Vuepress).
 
 **Developing the docs**
+
 ```shell
 npm run docs:dev
 ```
 
 **Building and serving the docs statically**
+
 ```shell
 npm run docs:build # Output docs to /dist
 npm run docs:serve # Serve docs from /dist
@@ -399,9 +401,9 @@ npm run test
 
 You can view a code coverage report generated in coverage/lcov-report/index.html.
 
-## Building faker.js
+## Building Faker
 
-faker uses [gulp](http://gulpjs.com/) to automate its build process. Each build operation is a separate task which can be run independently.
+Faker uses [gulp](http://gulpjs.com/) to automate its build process. Each build operation is a separate task which can be run independently.
 
 ### Browser Bundle
