+++ docs/.vitepress/theme/components/index.mjs
@@ -1,0 +1,7 @@
+import Playground from './Playground.vue';
+import Badge from './Badge.vue';
+
+export default {
+  Playground,
+  Badge,
+};
