--- docs/.vitepress/theme/components/index.ts
@@ -1,7 +1,0 @@
-import Playground from './Playground.vue'
-import Badge from './Badge.vue'
-
-export default {
-  Playground,
-  Badge,
-}
