+++ docs/.vitepress/theme/index.mjs
@@ -1,11 +1,11 @@
-import GlobalComponents from './components'
-import DefaultTheme from 'vitepress/theme'
+import GlobalComponents from './components';
+import DefaultTheme from 'vitepress/theme';
 
 export default {
   ...DefaultTheme,
   enhanceApp({ app }) {
     for (const [name, component] of Object.entries(GlobalComponents)) {
-      app.component(name, component)
+      app.component(name, component);
     }
-  }
-}
+  },
+};
