+++ docs/api/company.md
@@ -6,22 +6,23 @@
 
 Return random company name. Using one of the following format options
 ::: v-pre
+
 - `{{name.lastName}} {{company.companySuffix}}`
 - `{{name.lastName}} - {{name.lastName}}`
 - `{{name.lastName}}, {{name.lastName}} and {{name.lastName}}`
-:::
+  :::
 
 ::: tip
-| Param       | Type   | Default |
+| Param | Type | Default |
 | ----------- | ------ | :-----: |
-| formatIndex | number |  `0-2`  |
+| formatIndex | number | `0-2` |
 :::
 
 ```js
-faker.company.companyName(); // Zboncak and Sons 
+faker.company.companyName(); // Zboncak and Sons
 faker.company.companyName(0); // Nikolaus Group
-faker.company.companyName(1); // Keeling - Lind 
-faker.company.companyName(2); // Swaniawski, Howe and Leffler 
+faker.company.companyName(1); // Keeling - Lind
+faker.company.companyName(2); // Swaniawski, Howe and Leffler
 ```
 
 ## Company Suffix
@@ -53,7 +54,7 @@ faker.company.catchPhrase(); // Team-oriented context-sensitive conglomeration
 Return random catch phrase adjective
 
 ```js
-faker.company.catchPhraseAdjective(); // Down-sized 
+faker.company.catchPhraseAdjective(); // Down-sized
 ```
 
 ## Catchphrase Descriptor
@@ -61,7 +62,7 @@ faker.company.catchPhraseAdjective(); // Down-sized
 Return random catch phrase descriptor
 
 ```js
-faker.company.catchPhraseDescriptor(); // bi-directional 
+faker.company.catchPhraseDescriptor(); // bi-directional
 ```
 
 ## Catchphrase Noun
