+++ docs/api/database.md
@@ -3,11 +3,13 @@
 [[toc]]
 
 ## Column
+
 Return database column name
 
 ```js
 faker.database.column(); // createdAt
 ```
+
 ## Type
 
 Return database column type
@@ -21,7 +23,7 @@ faker.database.type(); // text
 Return database collation
 
 ```js
-faker.database.collation(); // cp1250_general_ci 
+faker.database.collation(); // cp1250_general_ci
 ```
 
 ## Engine
