+++ docs/api/date.md
@@ -7,20 +7,20 @@
 Return date sometime in the past.
 
 ::: tip
-| Param    | Type    |   Default    |
+| Param | Type | Default |
 | -------- | ------- | :----------: |
-| maxYears | number |     `1`      |
-| refDate  | Date    | `new Date()` |
+| maxYears | number | `1` |
+| refDate | Date | `new Date()` |
 :::
 
 ```js
-faker.date.past(); 
+faker.date.past();
 // Sat Oct 20 2018 04:19:38 GMT-0700 (Pacific Daylight Time)
 
 faker.date.past(100);
 // Tue Jul 27 1971 21:08:49 GMT-0700 (Pacific Daylight Time)
 
-faker.date.past(100, new Date(-3000)); 
+faker.date.past(100, new Date(-3000));
 // Wed Dec 07 1881 05:04:18 GMT-0752 (Pacific Standard Time)
 ```
 
@@ -29,21 +29,21 @@ faker.date.past(100, new Date(-3000));
 Return date sometime in the future.
 
 ::: tip
-| Param   | Type    |   Default    |
+| Param | Type | Default |
 | ------- | ------- | :----------: |
-| years   | number |     `1`      |
-| refDate | Date    | `new Date()` |
+| years | number | `1` |
+| refDate | Date | `new Date()` |
 :::
 
 ```js
 faker.date.future();
- // Mon Sep 02 2019 21:08:33 GMT-0700 (Pacific Daylight Time)
+// Mon Sep 02 2019 21:08:33 GMT-0700 (Pacific Daylight Time)
 
 faker.date.future(100);
- // Fri Nov 23 2068 10:43:15 GMT-0800 (Pacific Standard Time)
+// Fri Nov 23 2068 10:43:15 GMT-0800 (Pacific Standard Time)
 
 faker.date.future(100, new Date(-3000));
- // Mon Jun 26 2051 01:50:51 GMT-0700 (Pacific Daylight Time)
+// Mon Jun 26 2051 01:50:51 GMT-0700 (Pacific Daylight Time)
 ```
 
 ## Between
@@ -51,15 +51,15 @@ faker.date.future(100, new Date(-3000));
 Return date sometime in the future.
 
 ::: tip
-| Param | Type |                      Default                       |
+| Param | Type | Default |
 | ----- | ---- | :------------------------------------------------: |
-| from  | Date | <Badge text="N/A" type="error" vertical="middle"/> |
-| to    | Date | <Badge text="N/A" type="error" vertical="middle"/> |
+| from | Date | <Badge text="N/A" type="error" vertical="middle"/> |
+| to | Date | <Badge text="N/A" type="error" vertical="middle"/> |
 :::
 
 ```js
 faker.date.between(new Date(0), new Date(365 * 24 * 3600 * 1000));
- // Mon Oct 19 1970 14:12:01 GMT-0700 (Pacific Daylight Time)
+// Mon Oct 19 1970 14:12:01 GMT-0700 (Pacific Daylight Time)
 ```
 
 ## Recent <Badge text="5.0.0+" type="tip" vertical="middle"/>
@@ -67,10 +67,10 @@ faker.date.between(new Date(0), new Date(365 * 24 * 3600 * 1000));
 Returns some date recent date
 
 ::: tip
-| Param   | Type    |   Default    |
+| Param | Type | Default |
 | ------- | ------- | :----------: |
-| days    | number |     `1`      |
-| refDate | Date    | `new Date()` |
+| days | number | `1` |
+| refDate | Date | `new Date()` |
 :::
 
 ```js
@@ -82,10 +82,10 @@ faker.date.recent();
 Return date sometime soon.
 
 ::: tip
-| Param   | Type    |   Default    |
+| Param | Type | Default |
 | ------- | ------- | :----------: |
-| days    | number |     `1`      |
-| refDate | Date    | `new Date()` |
+| days | number | `1` |
+| refDate | Date | `new Date()` |
 :::
 
 ```js
@@ -97,7 +97,7 @@ faker.date.soon();
 Return random month name
 
 ::: tip
-| Param  | Type   |             Default             |
+| Param | Type | Default |
 | ------ | ------ | :-----------------------------: |
 | option | object | `{abbr: false, context: false}` |
 :::
@@ -106,7 +106,7 @@ Return random month name
 faker.date.month();
 // December
 
-faker.date.month({abbr: true});
+faker.date.month({ abbr: true });
 // Dec
 ```
 
@@ -115,15 +115,15 @@ faker.date.month({abbr: true});
 Return random weekday name
 
 ::: tip
-| Param  | Type   |             Default             |
+| Param | Type | Default |
 | ------ | ------ | :-----------------------------: |
 | option | object | `{abbr: false, context: false}` |
 :::
 
 ```js
 faker.date.weekday();
- // Sunday
+// Sunday
 
-faker.date.weekday({abbr: true});
- // Sun
+faker.date.weekday({ abbr: true });
+// Sun
 ```
