+++ docs/api/finance.md
@@ -7,9 +7,9 @@
 Return random account number
 
 ::: tip
-| Param  | Type   | Default |
+| Param | Type | Default |
 | ------ | ------ | :-----: |
-| length | number |   `8`   |
+| length | number | `8` |
 :::
 
 ```js
@@ -22,7 +22,7 @@ faker.finance.account(12); // 194294967472
 Return random account name
 
 ```js
-faker.finance.accountName(); // Investment Account 
+faker.finance.accountName(); // Investment Account
 ```
 
 ## Routing Number <Badge text="5.0.0+" type="tip" vertical="middle"/>
@@ -38,16 +38,16 @@ faker.finance.routingNumber();
 Return random mask
 
 ::: tip
-| Param    | Type    | Default |
+| Param | Type | Default |
 | -------- | ------- | :-----: |
-| length   | number  |   `4`   |
-| parens   | boolean | `false` |
+| length | number | `4` |
+| parens | boolean | `false` |
 | ellipsis | boolean | `false` |
 :::
 
 ```js
 faker.finance.mask(); // 8493
-faker.finance.mask(5, true); // (17126) 
+faker.finance.mask(5, true); // (17126)
 faker.finance.mask(5, true, true); // (...17126)
 faker.finance.mask(5, false, true); // ...17126
 ```
@@ -57,19 +57,19 @@ faker.finance.mask(5, false, true); // ...17126
 Return random amount
 
 ::: tip
-| Param  | Type   | Default |
+| Param | Type | Default |
 | ------ | ------ | :-----: |
-| min    | number |   `0`   |
-| max    | number | `1000`  |
-| dec    | number |   `2`   |
-| symbol | string |  `''`   |
+| min | number | `0` |
+| max | number | `1000` |
+| dec | number | `2` |
+| symbol | string | `''` |
 :::
 
 ```js
-faker.finance.amount(); // 466.98 
-faker.finance.amount(42); // 197.40  
+faker.finance.amount(); // 466.98
+faker.finance.amount(42); // 197.40
 faker.finance.amount(42, 43); // 42.24
-faker.finance.amount(42, 43, 10); // 42.1601538648 
+faker.finance.amount(42, 43, 10); // 42.1601538648
 faker.finance.amount(42, 43, 2, '$'); // $42.89
 ```
 
@@ -110,9 +110,9 @@ faker.finance.currencySymbol(); // £
 Return random credit card number
 
 ::: tip
-| Param    | Type   | Default |
+| Param | Type | Default |
 | -------- | ------ | :-----: |
-| provider | string |  `''`   |
+| provider | string | `''` |
 :::
 
 ```js
@@ -135,7 +135,6 @@ Return random bitcoin address
 faker.finance.bitcoinAddress(); // 3FINQD7C6JW7XPF4NMNHOJYAXLKLP
 ```
 
-
 ## Ethereum Address <Badge text="5.0.0+" type="tip" vertical="middle"/>
 
 Return random Ethereum address
@@ -149,7 +148,7 @@ faker.finance.ethereumAddress();
 Return random IBAN (International Bank Account Number)
 
 ::: tip
-| Param    | Type    | Default |
+| Param | Type | Default |
 | -------- | ------- | :-----: |
 | formated | boolean | `false` |
 :::
