+++ docs/guide/recent-faqs.md
@@ -2,13 +2,13 @@
 
 _January 11th, 2022_
 
-Due to recent events, folks may have some questions or concerns about the future of faker.js.
+Due to recent events, folks may have some questions or concerns about the future of Faker.
 
 Here are some answers.
 
 ## What happened?
 
-In short, the author deleted the original repository in protest against large companies who did not pay him for maintaining faker.js.
+In short, the author deleted the original repository in protest against large companies who did not pay him for maintaining Faker.
 
 In response to all that happened, we have created an entirely new GitHub Organization for the packages and ecosystem under the namespace [faker-js](https://github.com/faker-js).
 
@@ -19,33 +19,34 @@ In response to all that happened, we have created an entirely new GitHub Organiz
 Seriously, you don't have to sponsor _this_ project, just start having the conversation with your engineering leadership.
 :::
 
-
 ## Are these the official docs?
 
-**Yes**. This is the official site and documentation for the faker.js library.
+**Yes**. This is the official site and documentation for the Faker library.
 
-## Who maintains faker.js now?
+## Who maintains Faker now?
 
-**faker.js is now maintained by the organization [faker-js](https://github.com/faker-js)** after the original repository was deleted by the author.
-
+**Faker is now maintained by the organization [faker-js](https://github.com/faker-js)** after the original repository was deleted by the author.
 
 ## My build broke. Help.
 
-Due to the above issues, version 6 of faker.js **does not run**. Please pin your faker.js version to `5.5.3`.
+If you have installed `faker` from npm, please uninstall it and migrate to `@faker-js/faker`. Due to the above issues, version 6 of faker.js **does not run** and is why the `@faker-js/faker` package exists.
 
 We will be releasing a new version shortly.
 
 **NPM**
+
 ```shell
-npm install faker@5.5.3 --save-dev
+npm install @faker-js/faker --save-dev
 ```
 
 **Yarn**
+
 ```shell
-yarn add faker@5.5.3 -D
+yarn add @faker-js/faker -D
 ```
 
 **pnpm**
+
 ```shell
-pnpm add -D faker@5.5.3
+pnpm add -D @faker-js/faker
 ```
