+++ package.json
@@ -66,7 +66,6 @@
     "prettier": "2.5.1",
     "sinon": "^9.0.3",
     "through2": "2.0.0",
-    "typescript": "^4.5.4",
     "vinyl-buffer": "^1.0.1",
     "vinyl-source-stream": "^2.0.0",
     "vinyl-transform": "^1.0.0",
